services/api/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.35
asyncpg==0.29.0
httpx==0.27.2
redis==5.0.7
rq==1.16.2
jinja2==3.1.4
python-dotenv==1.0.1
alembic==1.13.2
prometheus-client==0.20.0
PyJWT==2.9.0
orjson==3.10.7
boto3==1.28.29
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.2
black==24.8.0
ruff==0.6.9
mypy==1.11.2
types-PyYAML==6.0.12.20240808
psycopg2-binary==2.9.9
python-multipart==0.0.9
